# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
      "ChecksumResponse": {
        "description": "Checksum metadata attached to stored content or versions.",
        "properties": {
          "algorithm": {
            "description": "Checksum algorithm used by the service.",
            "title": "Algorithm",
            "type": "string"
          },
          "digest": {
            "description": "Hex digest returned by the service.",
            "title": "Digest",
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "digest"
        ],
        "title": "ChecksumResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProvenanceResponse": {
        "description": "Minimal provenance returned by exact version reads.",
        "properties": {
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          },
          "tree_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tree Path"
          }
        },
        "required": [
          "repo_url",
          "commit_sha"
        ],
        "title": "ProvenanceResponse",
        "type": "object"
      },
...
      "RelationshipSelectorResponse": {
        "additionalProperties": false,
        "description": "Authored relationship selector preserved exactly as published.",
        "properties": {
          "markers": {
            "items": {
              "type": "string"
            },
            "title": "Markers",
            "type": "array"
          },
          "optional": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optional"
          },
          "slug": {
            "description": "Stable public slug of the related skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact related version when the selector targets one immutable version.",
            "title": "Version"
          },
          "version_constraint": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authored dependency constraint when the selector is version-ranged.",
            "title": "Version Constraint"
          }
        },
        "required": [
          "slug"
        ],
        "title": "RelationshipSelectorResponse",
        "type": "object"
      },
      "SkillContentSummaryResponse": {
        "description": "Compact content metadata returned without the full markdown body.",
        "properties": {
          "checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          },
          "rendered_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional pre-rendered short summary stored with the markdown.",
            "title": "Rendered Summary"
          },
          "size_bytes": {
            "description": "UTF-8 byte length of the stored markdown.",
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "checksum",
          "size_bytes",
          "rendered_summary"
        ],
        "title": "SkillContentSummaryResponse",
        "type": "object"
      },
...
      "SkillMetadataResponse": {
        "description": "Full normalized metadata block returned by exact fetch responses.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "inputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputs Schema"
          },
          "maturity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maturity Score"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Schema"
          },
          "security_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Score"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "token_estimate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Estimate"
          }
        },
        "required": [
          "name",
          "description",
          "tags"
        ],
        "title": "SkillMetadataResponse",
        "type": "object"
      },
...
      "SkillRelationshipResponse": {
        "description": "One authored relationship plus optional exact target enrichment.",
        "properties": {
          "selector": {
            "$ref": "#/components/schemas/RelationshipSelectorResponse"
          },
          "target_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillVersionReferenceResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "selector"
        ],
        "title": "SkillRelationshipResponse",
        "type": "object"
      },
...
      "SkillVersionReferenceResponse": {
        "description": "Compact exact version reference.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "name",
          "description",
          "tags",
          "lifecycle_status",
          "trust_tier",
          "published_at"
        ],
        "title": "SkillVersionReferenceResponse",
        "type": "object"
      },
...
      "SkillVersionRelationshipsResponse": {
        "description": "Grouped relationships returned in exact fetch responses.",
        "properties": {
          "conflicts_with": {
            "items": {
              "$ref": "#/components/schemas/SkillRelationshipResponse"
            },
            "title": "Conflicts With",
            "type": "array"
          },
          "depends_on": {
            "items": {
              "$ref": "#/components/schemas/SkillRelationshipResponse"
            },
            "title": "Depends On",
            "type": "array"
          },
          "extends": {
            "items": {
              "$ref": "#/components/schemas/SkillRelationshipResponse"
            },
            "title": "Extends",
            "type": "array"
          },
          "overlaps_with": {
            "items": {
              "$ref": "#/components/schemas/SkillRelationshipResponse"
            },
            "title": "Overlaps With",
            "type": "array"
          }
        },
        "title": "SkillVersionRelationshipsResponse",
        "type": "object"
      },
      "SkillVersionResponse": {
        "description": "Normalized exact immutable version metadata response.",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SkillContentSummaryResponse"
          },
          "content_download_path": {
            "title": "Content Download Path",
            "type": "string"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/SkillMetadataResponse"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "relationships": {
            "$ref": "#/components/schemas/SkillVersionRelationshipsResponse"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "version_checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          }
        },
        "required": [
          "slug",
          "version",
          "version_checksum",
          "content",
          "metadata",
          "lifecycle_status",
          "trust_tier",
          "relationships",
          "published_at",
          "content_download_path"
        ],
        "title": "SkillVersionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/skills/{slug}/versions/{version}": {
      "get": {
        "description": "Return the normalized metadata for one exact immutable skill version. This route does not inline the raw markdown body.",
        "operationId": "getExactSkillVersionMetadata",
        "parameters": [
          {
            "description": "Stable public slug of the skill to fetch.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill to fetch.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to fetch.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to fetch.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "content": {
                    "checksum": {
                      "algorithm": "sha256",
                      "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                    },
                    "rendered_summary": "Lint Python files consistently.",
                    "size_bytes": 123
                  },
                  "content_download_path": "/skills/python.lint/versions/1.2.3/content",
                  "lifecycle_status": "published",
                  "metadata": {
                    "description": "Linting skill",
                    "headers": {
                      "runtime": "python"
                    },
                    "inputs_schema": {
                      "type": "object"
                    },
                    "maturity_score": 0.9,
                    "name": "Python Lint",
                    "outputs_schema": {
                      "type": "object"
                    },
                    "security_score": 0.95,
                    "tags": [
                      "python",
                      "lint"
                    ],
                    "token_estimate": 128
                  },
                  "provenance": {
                    "commit_sha": "aabbccddeeff00112233445566778899aabbccdd",
                    "repo_url": "https://github.com/example/skills",
                    "tree_path": "skills/python.lint"
                  },
                  "published_at": "2026-03-10T08:30:00Z",
                  "relationships": {
                    "conflicts_with": [],
                    "depends_on": [
                      {
                        "selector": {
                          "markers": [
                            "linux",
                            "gpu"
                          ],
                          "optional": true,
                          "slug": "python.base",
                          "version_constraint": ">=1.0.0,<2.0.0"
                        }
                      }
                    ],
                    "extends": [
                      {
                        "selector": {
                          "markers": [],
                          "slug": "python.base",
                          "version": "1.0.0"
                        },
                        "target_version": {
                          "description": "Base runtime skill",
                          "lifecycle_status": "published",
                          "name": "Python Base",
                          "published_at": "2026-03-01T08:30:00Z",
                          "slug": "python.base",
                          "tags": [
                            "python",
                            "runtime"
                          ],
                          "trust_tier": "verified",
                          "version": "1.0.0"
                        }
                      }
                    ],
                    "overlaps_with": [
                      {
                        "selector": {
                          "markers": [],
                          "slug": "python.format",
                          "version": "1.0.0"
                        }
                      }
                    ]
                  },
                  "slug": "python.lint",
                  "trust_tier": "internal",
                  "version": "1.2.3",
                  "version_checksum": {
                    "algorithm": "sha256",
                    "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionResponse"
                }
              }
            },
            "description": "Exact immutable skill version metadata returned successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "slug": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body or path parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch exact immutable version metadata",
        "tags": [
          "fetch"
        ]
      }
    },
    "/skills/{slug}/versions/{version}/content": {
      "get": {
        "description": "Return the canonical markdown body for one exact immutable skill version with immutable cache headers.",
        "operationId": "getExactSkillVersionContent",
        "parameters": [
          {
            "description": "Stable public slug of the skill to fetch.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill to fetch.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to fetch.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to fetch.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/markdown": {}
            },
            "description": "Markdown content returned successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "slug": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body or path parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch exact immutable markdown content",
        "tags": [
          "fetch"
        ]
      }
    },
...
  }
}